app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
psycopg[binary]>=3.1
pgvector
PyMuPDF
google-genai
langchain-text-splitters
pydantic
